+++ GrizzCompany-Items/src/Plugin.cs
@@ -12,7 +12,7 @@ namespace GrizzCompany.Items
     {
         public const string GUID = "SylviBlossom.GrizzCompany-Items";
         public const string NAME = "GrizzCompany-Items";
-        public const string VERSION = "1.0.3";
+        public const string VERSION = "1.0.4";
 
         public static Plugin Instance { get; private set; }
 		public static new Config Config { get; private set; }